post test 진행중

+++ src/main/java/com/project/trip/post/controller/PostController.java
@@ -2,10 +2,12 @@ package com.project.trip.post.controller;
 
 import com.project.trip.global.oauth.CustomOauthUser;
 import com.project.trip.image.service.PostImageServiceImpl;
+import com.project.trip.post.entity.PostKind;
 import com.project.trip.post.model.request.PostSaveAndUpdateRequestDto;
+import com.project.trip.post.model.response.BoardResponseDto;
 import com.project.trip.post.model.response.PostResponseDto;
-import com.project.trip.post.model.response.PostSimpleResponseDto;
 import com.project.trip.post.service.PostServiceImpl;
+import com.project.trip.user.entity.Role;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.web.PageableDefault;
@@ -34,24 +36,46 @@ public class PostController {
 
     @Secured({"ROLE_USER", "ROLE_ADMIN"})
     @PostMapping("/posts")
-    public void save(@ModelAttribute("postSaveRequest") PostSaveAndUpdateRequestDto postSaveAndUpdateRequestDto,
-                     @RequestParam(value = "images") List<MultipartFile> images,
-                     @AuthenticationPrincipal CustomOauthUser oauthUser) throws IOException {
+    public ResponseEntity<?> save(@ModelAttribute("postSaveRequest") PostSaveAndUpdateRequestDto postSaveAndUpdateRequestDto,
+                                  @RequestParam(value = "images") List<MultipartFile> images,
+                                  @AuthenticationPrincipal CustomOauthUser oauthUser) throws IOException {
+        PostKind postKind = PostKind.convertFromString(postSaveAndUpdateRequestDto.getKind());
+
+        //TODO 하드코딩 리펙토링 고려
+        if (postKind == PostKind.NOTICE && oauthUser.getRole() == Role.USER) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         Long postId = postService.save(postSaveAndUpdateRequestDto, oauthUser.getEmail());
 
         postImageService.saveImage(images, postId);
+
+        return ResponseEntity.status(HttpStatus.OK).build();
     }
 
+    @Secured({"ROLE_USER", "ROLE_ADMIN"})
     @DeleteMapping("/posts/{postId}")
-    public void delete(@PathVariable Long postId) {
+    public ResponseEntity<?> delete(@PathVariable Long postId,
+                                    @AuthenticationPrincipal CustomOauthUser oauthUser) throws IllegalArgumentException{
+        boolean hasAuth = postService.checkAuth(oauthUser, postId);
+        System.out.println(hasAuth);
+
+        if (!hasAuth) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+        }
+
         postService.delete(postId);
+        return ResponseEntity.status(HttpStatus.OK).build();
     }
 
+    @Secured({"ROLE_USER", "ROLE_ADMIN"})
     @PutMapping("/posts/{postId}")
-    public void update(@RequestBody PostSaveAndUpdateRequestDto postUpdateRequestDto, @PathVariable Long postId) {
+    public void update(@RequestBody PostSaveAndUpdateRequestDto postUpdateRequestDto,
+                       @PathVariable Long postId) {
         postService.update(postUpdateRequestDto, postId);
     }
 
+    @Secured({"ROLE_USER", "ROLE_ADMIN"})
     @PostMapping("/posts/{postsId}/like")
     public void like(@PathVariable Long postsId,
                      @AuthenticationPrincipal CustomOauthUser oauthUser) {
@@ -60,21 +84,22 @@ public class PostController {
 
     @GetMapping("/posts/{postId}")
     public PostResponseDto getPostById(@PathVariable Long postId) {
-
         return postService.getPostDtoById(postId);
     }
 
     @GetMapping("/board")
-    public void getBoard(@PathVariable String boardKind,
-                         @PageableDefault(size = 5) Pageable pageable) {
-        List<PostSimpleResponseDto> noticeList = postService.getNotices();
-        List<PostSimpleResponseDto> postList = postService.getPostsByKind(pageable);
+    public BoardResponseDto getBoard(@PathVariable String boardKind,
+                                     @PageableDefault(size = 5) Pageable pageable) {
+        BoardResponseDto boardResponseDto = new BoardResponseDto();
 
-        //TODO 위 두 List를 어떻게 반환할지 고민 return
+        boardResponseDto.setNoticeList(postService.getNotices());
+        boardResponseDto.setPostList(postService.getPostsByKind(boardKind, pageable));
+
+        return boardResponseDto;
     }
 
     @ExceptionHandler(IllegalArgumentException.class)
-    public ResponseEntity<?> handleMethodNoPostFoundException(MethodArgumentTypeMismatchException ex) {
+    public ResponseEntity<?> handleMethodNoPostFoundException(IllegalArgumentException ex) {
         return ResponseEntity.status(HttpStatus.BAD_REQUEST)
                 .body("no posts founded");
     }
